# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend created by Hanif using Gin",
        "title": "Backend Ewallet",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/user/password": {
            "patch": {
                "description": "Change the current user password to new password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update User Password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Password Update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdatePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/pin": {
            "patch": {
                "description": "Change or set new security PIN for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update User PIN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "PIN Update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdatePinReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "description": "retrieved detailed information of currently logged in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current user Profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "\u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserProfileRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update detailed information of the currently logged in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update User Profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Profile Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdateProfileReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserUpdateProfilRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/profile/pin/check": {
            "post": {
                "description": "Verify if the provided pin is correct for the logged in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Check pin user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "PIN payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserCheckPinReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserCheckPinRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/reports": {
            "get": {
                "description": "Retrieve transaction reports with duration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get User Transaction Report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "7d",
                        "description": "Duration of report, default is 7d",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-array_github_com_iamhanif11_ewallet-backend_internal_dto_UserTransactionReportRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/wallet": {
            "get": {
                "description": "Retrieve data for the user dashboard",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user Dashboard information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003ctoken\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserDashboardInformationRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_iamhanif11_ewallet-backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-array_github_com_iamhanif11_ewallet-backend_internal_dto_UserTransactionReportRes": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserTransactionReportRes"
                    }
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserCheckPinRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserCheckPinRes"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserDashboardInformationRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserDashboardInformationRes"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserProfileRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserProfileRes"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.Response-github_com_iamhanif11_ewallet-backend_internal_dto_UserUpdateProfilRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdateProfilRes"
                },
                "isSucces": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserCheckPinReq": {
            "type": "object",
            "required": [
                "pin"
            ],
            "properties": {
                "pin": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserCheckPinRes": {
            "type": "object",
            "properties": {
                "isvalid": {
                    "type": "boolean"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserDashboardInformationRes": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "expense": {
                    "type": "integer"
                },
                "income": {
                    "type": "integer"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserProfileRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserTransactionReportRes": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                },
                "expense": {
                    "type": "integer"
                },
                "income": {
                    "type": "integer"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdatePasswordReq": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdatePinReq": {
            "type": "object",
            "required": [
                "pin"
            ],
            "properties": {
                "pin": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdateProfilRes": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        },
        "github_com_iamhanif11_ewallet-backend_internal_dto.UserUpdateProfileReq": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token used for authorization",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}